Session 7 finished

+++ src/main/java/algstudent/s7/util/Heap.java
@@ -58,7 +58,7 @@ public class Heap {
 	 * Gets the heuristic of the best node in the priority queue
 	 * @return Value of the heuristic of the best node in the queue
 	 */
-	public int estimateBest() {
+	public double estimateBest() {
 		return nodes.peek().getHeuristicValue();
 	}
